Move point queries to graph
