+++ src/fpm/graph.py
@@ -7,7 +7,7 @@ import rdflib
 from rdflib import RDF
 
 from fpm import traversal
-from fpm.constants import GEO, GEOM, COORD, COORD_EXT, QUDT, QUDT_VOCAB, FP
+from fpm.constants import GEO, GEOM, COORD, COORD_EXT, QUDT, QUDT_VOCAB, FP, POLY
 from fpm.utils import build_transformation_matrix
 
 
@@ -149,3 +149,36 @@ def get_transformation_matrix_wrt_frame(g, root, target):
         T = np.dot(new_T, T)
 
     return T
+
+
+def get_space_points(g):
+    floorplan = g.value(predicate=RDF.type, object=FP["FloorPlan"])
+
+    # Get the list of spaces
+    print("Querying all spaces...")
+    space_ptr = g.value(floorplan, FP["spaces"])
+    spaces = get_list_from_ptr(g, space_ptr)
+
+    # for each space, find the polygon
+    print("Get all points...")
+    space_points = []
+    for space in spaces:
+        space_points_json = get_point_positions_in_space(g, space)
+        space_points.append(space_points_json)
+
+    return space_points
+
+
+def get_point_positions_in_space(g, space):
+    polygon = g.value(space, FP["shape"])
+
+    point_ptr = g.value(polygon, POLY["points"])
+
+    point_nodes = get_list_from_ptr(g, point_ptr)
+
+    positions = []
+    for point in point_nodes:
+        position = get_point_position(g, point)
+        positions.append(position)
+
+    return {"name": prefixed(g, space), "points": positions}
